Use PaymentMethod instead of BillingIntegration
